SmartPDF-OCR/requirements.txt
# SmartPDF-OCR Dependencies
# Python 3.10+

# OCR Engine
paddlepaddle-gpu==2.5.2
paddleocr==2.7.3

# PDF Processing
pdfplumber==0.10.3
PyMuPDF==1.23.7

# Image Processing
opencv-python==4.8.1.78
numpy==1.26.2
Pillow==10.1.0

# Web Framework
fastapi==0.128.0
uvicorn[standard]==0.24.0
python-multipart==0.0.21

# Frontend
gradio==6.2.0

# Document Export
python-docx==1.1.0
reportlab==4.0.7

# Utilities
pydantic==2.12.5
pydantic-settings==2.1.0
aiofiles==23.2.1
